fix: address PR 370 review comments and simplify code quality (#377)
* fix: address PR 370 review comments

- layout.tsx: localize overflow menu strings (overflowMenuTriggerText /
  overflowMenuTitleText) via locale check so they switch between
  Japanese and English with the rest of the header
- aws/index.ts: remove duplicate error.message.includes("does not exist")
  condition that was a copy-paste error
- tenant-management/index.ts: paginate through all tenants instead of
  capping at 100; totalTenants and per-status counters now use the same
  full dataset, fixing incorrect uptime % and provisioningStats when
  tenant count exceeds 100


* simplify: extract shared localStorage mock, split problem-repository, interface fix

- Extract createLocalStorageMock to lib/__tests__/test-helpers.ts; remove
  inline copies from i18n.test.tsx and notifications/context.test.tsx
- Split problem-repository.ts (690 lines) into three focused modules:
  problem-mapper.ts (type converters), marketplace-repository.ts
  (PrismaMarketplaceRepository), problem-repository.ts (PrismaProblemRepository)
- Add findByExternalId to IProblemRepository interface and
  InMemoryProblemRepository; remove unsafe type cast in
  PrismaProblemRepository.findByExternalId
- gameday-deployer: inline single-use article ternary variable
- tenant-management index: single for-of loop replaces four Array.filter
  passes; full pagination so stats totalTenants reflects all tenants
- provisioning: return from inner catch instead of re-throwing to eliminate
  duplicate error log entries
- context.test.tsx: split 8-expect notification test into three focused tests


* ci: trigger CodeQL re-scan


---------
